websockety

+++ app/Services/UserService.php
@@ -7,6 +7,7 @@ namespace Sms\Services;
 use Carbon\Carbon;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Hash;
+use Sms\Events\Event;
 use Sms\Models\AgencyRole;
 use Sms\Models\User;
 
@@ -22,15 +23,13 @@ class UserService extends BaseService
      */
     public function create(array $request): User
     {
-        return User::create([
-            'name' => $request['name'],
-            'surname' => $request['surname'],
-            'phone_number' => $request['phone_number'],
-            'email' => $request['email'],
-            'password' => $this->hashPassword($request['password']),
-            'agency_role_id' => $request['agency_role_id'],
-            'agency_id' => $request['agency_id']
-        ]);
+        $user = new User($request);
+        $user->save();
+
+        event(new Event(["users"], 'update'));
+        event(new Event(["agency-$user->agency_id"], 'statistics'));
+
+        return $user;
     }
 
     /**
@@ -91,6 +90,9 @@ class UserService extends BaseService
 
         $user->save();
 
+        event(new Event(["user-$user->id"], 'update', ['user' => $user]));
+        event(new Event(["users"], 'update'));
+
         return $user;
     }
 
@@ -106,8 +108,12 @@ class UserService extends BaseService
             $user = auth()->user()->agency->employees()->findOrFail($userId);
         }
 
-        if (auth()->id() != $userId) {
+        if (auth()->id() != $userId && auth()->user()->isManager()) {
             $user->delete();
+
+            event(new Event(["users"], 'update'));
+            event(new Event(["agency-$user->agency_id"], 'statistics'));
+
             return true;
         }
 
@@ -122,6 +128,9 @@ class UserService extends BaseService
         $user = User::findOrFail($userId);
         $user->blocked_at = Carbon::now();
         $user->save();
+
+        event(new Event(["user-$user->id"], 'update'));
+        event(new Event(["users"], 'update'));
     }
 
     /**
@@ -132,6 +141,9 @@ class UserService extends BaseService
         $user = User::findOrFail($userId);
         $user->blocked_at = null;
         $user->save();
+
+        event(new Event(["user-$user->id"], 'update'));
+        event(new Event(["users"], 'update'));
     }
 
     /**
